Several additions
This commit fixes issue #3 and #4

Feature added: Read weather data from a text file.
Some modifications to the GUI. Addition of output window and functions
to communicate with the output window.

Structure of Current changed. Now carries an ID, so an order can be kept
on the currents in the system.

More logic added to the way balloons shift between currents. They move
up one or down one.  If in the lowest, the only change possible is
upwards. Likewise if in the top layer, only change allowed is down.

Global variables moved from class Window to class World where they can
be used by both classes.